elfloader: print bootloader_dtb instead of dtb
Minor fixup, the dtb variable is always 0 when printed in
sys_boot.c as it refers to the the address the dtb is loaded
to after load_images has finished loading it from either the
bootloader_dtb or cpio archive.

Print bootloader_dtb instead, since it contains the address
to a valid dtb given by the bootloader or efi.

Signed-off-by: Viktor Sannum <sannum.viktor@gmail.com>